student-work/Jaffer/facial-emotion-recognition/requirements.txt
numpy==1.24.3
pandas==2.1.3
opencv-python==4.8.1.78
pillow==10.3.0
streamlit==1.39.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.9.0
torch==.1.1
torchvision==0.16.1
tqdm==4.65.0
scikit-learn==1.3.0
clip==0.2.0
transformers==4.32.1